Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -9,10 +9,10 @@ use App\Http\Requests\AskQuestionRequest;
 class QuestionsController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    * Display a listing of the resource.
+    *
+    * @return \Illuminate\Http\Response
+    */
     public function index()
     {
         $questions = Question::with('user')->latest()->paginate(10);
@@ -21,10 +21,10 @@ class QuestionsController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    * Show the form for creating a new resource.
+    *
+    * @return \Illuminate\Http\Response
+    */
     public function create()
     {
         $question = new Question();
@@ -33,11 +33,11 @@ class QuestionsController extends Controller
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+    * Store a newly created resource in storage.
+    *
+    * @param  \Illuminate\Http\Request  $request
+    * @return \Illuminate\Http\Response
+    */
     public function store(AskQuestionRequest $request)
     {
         $request->user()->questions()->create($request->only("title", "body"));
@@ -46,11 +46,11 @@ class QuestionsController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  \App\Question  $question
-     * @return \Illuminate\Http\Response
-     */
+    * Display the specified resource.
+    *
+    * @param  \App\Question  $question
+    * @return \Illuminate\Http\Response
+    */
     public function show(Question $question)
     {
         $question->increment("views");
@@ -59,38 +59,52 @@ class QuestionsController extends Controller
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Question  $question
-     * @return \Illuminate\Http\Response
-     */
+    * Show the form for editing the specified resource.
+    *
+    * @param  \App\Question  $question
+    * @return \Illuminate\Http\Response
+    */
     public function edit(Question $question)
     {
+        if (\Gate::denies("update-question", $question)) {
+            abort(403, "Access denied");
+        }
+
         return view("questions.edit", compact("question"));
+
+
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Question  $question
-     * @return \Illuminate\Http\Response
-     */
+    * Update the specified resource in storage.
+    *
+    * @param  \Illuminate\Http\Request  $request
+    * @param  \App\Question  $question
+    * @return \Illuminate\Http\Response
+    */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies("update-question", $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only("title", "body"));
 
         return redirect()->route("questions.index")->with("success", "Your question has been updated.");
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Question  $question
-     * @return \Illuminate\Http\Response
-     */
+    * Remove the specified resource from storage.
+    *
+    * @param  \App\Question  $question
+    * @return \Illuminate\Http\Response
+    */
     public function destroy(Question $question)
     {
+        if (\Gate::denies("delete-question", $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->delete();
 
         return redirect("/questions")->with("success", "Your question has been deleted.");
